Add details of main inverter wiring harness
This expands the rear inverter connector pinout to include the
wiring harness colours found on the Tesla harness (1067968-
02-E). Also, add details of the connector pin numbering and
physical layout. With these changes it is possible to test and
modify Tesla wiring harnesses more easily.

Add details of missing Stator Temp + and - signal.

Remove the PDF in favour of an ODS spreadsheet to allow
easier future revision.

Tests:
 - Cross check against Tesla wiring harness with a multimeter