Add an Arduino sketch to control the Tesla oil pump
This Arduino sketch connects to the oil pump over LIN using the
hardware serial on an ESP32. According to the captures taken of
the LIN of a running inverter there is 1 outgoing command and
4 status requests. The outgoing command appears to be the
desired speed of the pump. The incoming status messages are
currently not understood so are just printed out as hex values.

The speed can be changed by supplying a numeric value between
0 and 255 followed by CR/LF.