Decode oil pump operating status fields
Run the oil pump for about a minute at full speed and
capture the log in short-run.ods. This allows various
status parameters to be graphed against time.

It looks like 30_Byte3 ramps up first and the graph
follows the pitch which can be heard from the motor.
Label this as Motor_Speed.

The 2A_Byte1 values ramp up next and this reflects the
observed behaviour when emptying the old oil out of
my drive unit where there was a perceptible delay
between the motor spinning up and the fluid flow. This
value ramps down with a lag after the Speed_Request
returns to 0.

The 2A_Byte2 value looks like the best guess for the
pump pressure. It peaks a bit after the flow and RPM
have stabilised and tails away steadily. This would
likely be expected in my drive unit which only has a
small amount of oil it it currently. No idea if the value is
from a sensor or some sort of clever fusion from the 3-
phase motor controller.